修改查找评论 返回评论列表新增关联文章标题

+++ src/main/java/com/zsc/blog/service/impl/TCommentServiceImpl.java
@@ -1,6 +1,7 @@
 package com.zsc.blog.service.impl;
 
 import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.zsc.blog.Utils.RedisUtil;
 import com.zsc.blog.entity.TComment;
 import com.zsc.blog.entity.TUser;
 import com.zsc.blog.mapper.TCommentMapper;
@@ -9,7 +10,9 @@ import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.annotation.Resource;
 import java.util.List;
+import java.util.Map;
 
 /**
  * <p>
@@ -23,6 +26,9 @@ import java.util.List;
 public class TCommentServiceImpl extends ServiceImpl<TCommentMapper, TComment> implements ITCommentService {
     @Autowired
     TCommentMapper tCommentMapper;
+    @Resource
+    RedisUtil redisUtil;
+
     @Override
     public Object selectList(Object o) {
         return tCommentMapper.selectList(null);
@@ -34,19 +40,46 @@ public class TCommentServiceImpl extends ServiceImpl<TCommentMapper, TComment> i
     }
 
     @Override
-    public void insertcomment(TComment tComment) {
-        tCommentMapper.insert(tComment);
+    public void insertcomment(TComment tComment)
+    {
+        String username=tComment.getAuthor();
+        if(redisUtil.get("commentlist_"+username)==null) {
+            tCommentMapper.insert(tComment);
+        }
+        else
+        {
+            redisUtil.del("commentlist_"+username);
+        }
         System.out.println("插入评论成功！");
     }
 
     @Override
-    public List<TComment> Selectbyuid(String username) {
-        return tCommentMapper.selectList(new QueryWrapper<TComment>().eq("author", username));
+    public List<Map<String,String>> Selectbyuid(String username) {
+        List<Map<String,String>> CommentList;
+        if (redisUtil.get(  "commentlist_"+username) == null)
+        {
+            CommentList = tCommentMapper.selcetcommentbyusername(username);
+            redisUtil.set("commentlist_"+username,CommentList,18000);
+
+        }
+        else
+        {
+            CommentList=(List<Map<String,String>>) redisUtil.get("commentlist_"+username);
+        }
+        return CommentList;
     }
 
     @Override
-    public void Delcomment(int aid, String username) {
-        tCommentMapper.deleteCommentWithauthor(aid,username);
+    public void Delcomment(int aid, String username)
+    {
+        if(redisUtil.get("commentlist_"+username)==null)
+        {
+            tCommentMapper.deleteCommentWithauthor(aid, username);
+        }
+        else
+        {
+            redisUtil.del("commentlist_"+username);
+        }
     }
 
     @Override
